Refactor (user) : refactor the codes based on reviewer's comments
