+++ src/main/java/atdd/Constant.java
@@ -1,4 +1,4 @@
-package atdd.path.web;
+package atdd;
 
 public class Constant {
     public static final String USER_BASE_URI = "/users";
@@ -6,4 +6,5 @@ public class Constant {
     public static final String PATH_BASE_URI = "/paths";
     public static final String STATION_BASE_URI = "/stations";
     public static final String LOGIN_BASE_URI = "/login";
+    public static final String AUTH_SCHEME_BEARER = "Bearer ";
 }
