--- src/main/java/atdd/path/application/LoginService.java
@@ -1,4 +1,0 @@
-package atdd.path.application;
-
-public class LoginService {
-}
