+++ src/main/java/atdd/user/application/dto/LoginReponseView.java
@@ -1,4 +1,4 @@
-package atdd.path.application.dto;
+package atdd.user.application.dto;
 
 public class LoginReponseView {
     private String accessToken;
