+++ src/main/java/atdd/user/application/exception/InvalidJwtAuthenticationException.java
@@ -1,4 +1,4 @@
-package atdd.path.application.exception;
+package atdd.user.application.exception;
 
 public class InvalidJwtAuthenticationException extends RuntimeException  {
     public InvalidJwtAuthenticationException(String message) {
