+++ src/main/java/atdd/user/domain/UserRepository.java
@@ -1,4 +1,4 @@
-package atdd.path.domain;
+package atdd.user.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
