+++ src/main/java/atdd/user/web/UserController.java
@@ -1,9 +1,9 @@
-package atdd.path.web;
+package atdd.user.web;
 
-import atdd.path.application.UserService;
-import atdd.path.application.dto.CreateUserRequestView;
-import atdd.path.application.dto.UserResponseView;
-import atdd.path.domain.User;
+import atdd.user.application.UserService;
+import atdd.user.application.dto.CreateUserRequestView;
+import atdd.user.application.dto.UserResponseView;
+import atdd.user.domain.User;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -14,7 +14,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.net.URI;
 
-import static atdd.path.web.Constant.USER_BASE_URI;
+import static atdd.Constant.USER_BASE_URI;
 
 @RestController
 @RequestMapping(USER_BASE_URI)
