+++ src/test/java/atdd/user/jwt/JwtTokenProviderTest.java
@@ -1,11 +1,12 @@
-package atdd.path.jwt;
+package atdd.user.jwt;
 
+import atdd.user.jwt.JwtTokenProvider;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+@SpringBootTest(classes = JwtTokenProvider.class)
 public class JwtTokenProviderTest {
     @Autowired
     private JwtTokenProvider jwtTokenProvider;
