+++ src/test/java/atdd/user/jwt/ReadPropertiesTest.java
@@ -1,5 +1,6 @@
-package atdd.path.jwt;
+package atdd.user.jwt;
 
+import atdd.user.jwt.Properties;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
